// src/modules/roles/constants/roles.enums.ts
export enum RolesErrorKeyEnum {
  FAILED_TO_FETCH_ROLES = 'Failed to fetch roles',
  ROLE_IS_REQUIRED = 'Role is required',
  FAILED_TO_CREATE_ROLE = 'Failed to create role',
  ROLE_ID_ARE_REQUIRED = 'Role ID  are required',
  FAILED_TO_UPDATE_ROLE = 'Failed to update role',
  BOTH_ROLE_ID_ARE_REQUIRED = 'Both Role  ID are required',
  FAILED_TO_REMOVE_ROLE = 'Failed to remove role',
  FAILED_TO_RETRIEVE_ROLE_PERMISSIONS = 'Failed to retrieve role permissions',
  BOTH_ROLE_ID_AND_PERMISSION_ID_ARE_REQUIRED = 'Both Role ID and Permission ID are required',
  FAILED_TO_ADD_PERMISSION_TO_ROLE = 'Failed to add permission to role',
  FAILED_TO_REMOVE_PERMISSION_FROM_ROLE = 'Failed to remove permission from role',
}